+++ src/main/java/net/silentchaos512/scalinghealth/ScalingHealth.java
@@ -1,10 +1,13 @@
 package net.silentchaos512.scalinghealth;
 
+import net.minecraft.item.ItemGroup;
+import net.minecraft.item.ItemStack;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.fml.DistExecutor;
 import net.minecraftforge.fml.ModContainer;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.common.Mod;
+import net.silentchaos512.scalinghealth.init.ModItems;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -15,12 +18,19 @@ import java.util.Random;
 public class ScalingHealth {
     public static final String MOD_ID = "scalinghealth";
     public static final String MOD_NAME = "Scaling Health";
-    public static final String VERSION = "2.3.6";
+    public static final String VERSION = "2.3.7";
     public static final String RESOURCE_PREFIX = MOD_ID + ":";
 
     public static final Random random = new Random();
 
     private static ScalingHealth INSTANCE;
+    public static final ItemGroup SH = new ItemGroup("scalinghealth") {
+        @Override
+        public ItemStack createIcon() {
+            return new ItemStack(ModItems.HEART_CRYSTAL.asItem());
+        }
+    };
+
     public static SideProxy PROXY;
 
     public static final Logger LOGGER = LogManager.getLogger(MOD_NAME);