+++ src/main/java/net/silentchaos512/scalinghealth/client/particles/ModParticles.java
@@ -1,17 +1,26 @@
 package net.silentchaos512.scalinghealth.client.particles;
 
+import net.minecraft.client.Minecraft;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.particles.BasicParticleType;
 import net.minecraft.particles.ParticleType;
+import net.minecraft.particles.ParticleTypes;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.world.IWorld;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.event.ParticleFactoryRegisterEvent;
 import net.minecraftforge.event.RegistryEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.silentchaos512.scalinghealth.ScalingHealth;
 import net.silentchaos512.utils.Color;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Locale;
 
+@Mod.EventBusSubscriber(modid = ScalingHealth.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD)
 public enum ModParticles {
     HEART_CRYSTAL("heart_crystal", Color.FIREBRICK, 0.08, 0.05),
     POWER_CRYSTAL("power_crystal", Color.ROYALBLUE, 0.08, 0.05),
@@ -45,14 +54,22 @@ public enum ModParticles {
         }
     }
 
-//    @OnlyIn(Dist.CLIENT)
+    @SubscribeEvent
     public static void registerAll(RegistryEvent.Register<ParticleType<?>> event) {
-        // FIXME: This does not work AT ALL. As far as I can tell, we need to register an
-        //  IParticleFactory, but all the methods are private.
         for (ModParticles particle : values()) {
             ResourceLocation id = ScalingHealth.getId(particle.name().toLowerCase(Locale.ROOT));
             particle.particleType.setRegistryName(id);
             ForgeRegistries.PARTICLE_TYPES.register(particle.particleType);
+        }
+    }
+
+    @Mod.EventBusSubscriber(value = Dist.CLIENT, modid = ScalingHealth.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD)
+    public static class FactoryHandler{
+        @SubscribeEvent
+        public static void registerFactories(ParticleFactoryRegisterEvent event){
+            for(ModParticles particle : values()){
+                Minecraft.getInstance().particles.registerFactory(particle.particleType, new ModParticle.Factory(particle.color));
+            }
         }
     }
 }