+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.7] - 2020-01-10
+### Added
+- Power Crystal now has a texture!
+- Scaling Health now has a creative tab
+
 ## [2.3.6] - 2020-01-09
 ### Fixed
 - Blight's fire now render! Hooray!